Move classes to their own files
No logic was changed.  Simply separated out each class into its own file.
