+++ src/System.Net.WebSockets.Client/tests/System.Net.WebSockets.Client.Tests.csproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="12.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003" DefaultTargets="Build">
   <PropertyGroup>
     <Configuration Condition="'$(Configuration)'==''">Windows_Debug</Configuration>
@@ -29,9 +29,15 @@
     <Compile Include="$(CommonTestPath)\System\Net\WebSocketTestServers.cs">
       <Link>Common\System\Net\WebSocketTestServers.cs</Link>
     </Compile>
-    <Compile Include="ClientWebSocketTest.cs" />
+    <Compile Include="AbortTest.cs" />
+    <Compile Include="CancelTest.cs" />
+    <Compile Include="ClientWebSocketTestBase.cs" />
     <Compile Include="ClientWebSocketUnitTest.cs" />
+    <Compile Include="CloseTest.cs" />
+    <Compile Include="ConnectTest.cs" />
+    <Compile Include="KeepAliveTest.cs" />
     <Compile Include="ResourceHelper.cs" />
+    <Compile Include="SendReceiveTest.cs" />
     <Compile Include="WebSocketData.cs" />
     <Compile Include="WebSocketHelper.cs" />
   </ItemGroup>
